rename consumer-bit-json.js to consumer-bit-config.js because the config is going to be written in package.json as well
