+++ src/consumer/bit-json/consumer-bit-config.js
